coding-challenge-2/gondo/apps/notification-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-json-logger>=3.0.0
pytest>=8.0.0
ruff>=0.8.0
httpx>=0.28.0
grpcio>=1.68.0
grpcio-tools>=1.68.0
protobuf>=5.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
asyncpg>=0.30.0
nats-py>=2.7.0
